Company2

+++ app/Http/Controllers/Company/ArtistController.php
@@ -43,8 +43,11 @@ use Intervention\Image\ImageManagerStatic as Image;
 class ArtistController extends Controller
 {
 
-    public function index()
+    public function index(Request $request)
     {
+        if(!$request->hasValidSignature()){
+            return abort(401);
+        }
         Permit::where('created_by', Auth::user()->user_id)->update(['is_edit' => 0]);
         ArtistTempData::where('created_by', Auth::user()->user_id )->where('status' , 0)->delete();
         Permit::whereDate('expired_date', '<', Carbon::now())->update(['permit_status' => 'expired']);
@@ -53,14 +56,20 @@ class ArtistController extends Controller
 
     // Artist Permit Dashboard Table One
 
-    public function fetch_applied()
+    public function fetch_applied(Request $request)
     {
-        return $this->datatable_function('applied');
+        if($request->ajax())
+        {   
+            return $this->datatable_function('applied');
+        }
     }
 
-    public function fetch_valid()
+    public function fetch_valid(Request $request)
     {
-        return $this->datatable_function('valid');
+        if($request->ajax())
+        {
+            return $this->datatable_function('valid');
+        }
     }
 
     function datatable_function($status)
@@ -428,9 +437,9 @@ class ArtistController extends Controller
 
     public function create(Request $request, $id = null)
     {
-        // if(!$request->hasValidSignature()){
-        //     return abort(401);
-        // }
+        if(!$request->hasValidSignature()){
+            return abort(401);
+        }
         
         $last_page = URL::previous();
 
@@ -445,8 +454,11 @@ class ArtistController extends Controller
             $id = $existTemp->permit_id;
         }
 
-        $add_artist_url = url('company/artist/add_new/' . $id);
-        $add_permit_url = url('company/artist/new/' . $id);
+        // $add_artist_url = url('company/artist/add_new/' . $id);
+        // $add_permit_url = url('company/artist/new/' . $id);
+        
+        $add_artist_url = URL::signedRoute('company.add_new_artist' , [ 'id' => $id ]);
+        $add_permit_url = URL::signedRoute('company.add_new_permit' , [ 'id' => $id]);
 
         if ($add_permit_url != $last_page && $last_page != $add_artist_url && !$view_artist_url_check && !$edit_artist_url_check) {
             ArtistTempData::where('permit_id', 1)->where('status', '!=', 5)->where('created_by', Auth::user()->user_id)->delete();
@@ -560,6 +572,21 @@ class ArtistController extends Controller
         $documentDetails = json_decode($request->documentD, true);
         $date_time = date('d_m_Y_H_i_s');
         $permit_id = $request->permit_id;
+
+        $toURL = '';
+        try {
+            
+            if($request->btnOption == 1){
+                if($request->from  == 'draft'){
+                    $toURL = URL::signedRoute('company.view_draft_details', [ 'id' => $permit_id]) ;
+                }else if ($request->event == 'event') {
+                    $toURL = URL::signedRoute('event.add_artist', [ 'id' => $permit_id]) ;
+                }else {
+                    $toURL = URL::signedRoute('company.add_new_permit', [ 'id' => $permit_id]);
+                }
+            }else if($request->btnOption == 2) {
+                $toURL = URL::signedRoute('company.add_new_artist', ['id' => $permit_id]);
+            }
 
         $artistTempData  = ArtistTempData::create([
             'firstname_en' => $artistDetails['fname_en'],
@@ -735,14 +762,17 @@ class ArtistController extends Controller
             }
         }
 
-
-        if ($artistTempData) {
+            DB::commit();
+		
             $result = ['success', __('Artist Added Successfully'), 'Success'];
-        } else {
-            $result = ['error', __('Error, Please Try Again'), 'Error'];
+        
+        } catch (Exception $e) {
+            DB::rollBack();
+            
+            $result = ['error', __($e->getMessage()), 'Error'];
         }
 
-        return response()->json(['message' => $result]);
+        return response()->json(['message' => $result , 'toURL' => $toURL]);
     }
 
     public function clear_the_temp_data(Request $request)
@@ -779,13 +809,27 @@ class ArtistController extends Controller
 
     public function store(Request $request)
     {
+        $toURL = '';
+
         try {
+
             DB::beginTransaction();
         
         $temp_permit_id = $request->temp_permit_id;
         $user_id = Auth::user()->user_id;
         $date_time = date('d_m_Y_H_i_s');
         $currentDateTime = Carbon::now()->toDateTimeString();
+        $fromWhere = $request->fromWhere ;
+        if( $fromWhere == 'event')
+        {
+            $toURL = URL::signedRoute('event.index').'#applied';
+        }else if( $fromWhere == 'renew')
+        {
+            $toURL = URL::signedRoute('artist.index').'#valid';
+        }else 
+        {
+            $toURL =URL::signedRoute('artist.index').'#applied';
+        }
 
         $artist_temp_data = ArtistTempData::where([
             ['permit_id', $temp_permit_id],
@@ -904,9 +948,10 @@ class ArtistController extends Controller
 
                             $artist_id = $artistTable->artist_id;
 
-                          
+                            // $artistTable->companies()->sync($request->user()->company->company_id); 
+                            // $request->user()->company->artist()->sync($artist_id); 
+                            // $request->user()->company->artists()->where('artist_id', '!=', $artist_id)->attach($artist_id);
 
-                           
                         } else {
                             $artist_id = $data->artist_id;
                         }
@@ -919,6 +964,11 @@ class ArtistController extends Controller
                         $artistPermitId = $artistPermit->artist_permit_id;
 
                     }
+
+                    // if($request->user()->company->artists()->where('artist.artist_id', '!=', $artist_id)->exists()){
+                    //     $request->user()->company->artists()->attach($artist_id);
+                    // } 
+
                     $org = [];
 
                     if($data->original)
@@ -1072,7 +1122,7 @@ class ArtistController extends Controller
         }
 
 
-        return response()->json(['message' => $result]);
+        return response()->json(['message' => $result, 'toURL' => $toURL ]);
     }
 
     public function clear_the_temp()
@@ -1259,13 +1309,20 @@ class ArtistController extends Controller
 
     public function delete_artist_from_temp(Request $request)
     {
-        $del_temp_id  = $request->del_temp_id;
-        $permit_id = $request->del_permit_id;
-        ArtistTempData::where('id', $del_temp_id)->update(['del_status' => 1]);
-        // ArtistTempData::where('id', $del_temp_id)->delete();
-        // ArtistTempDocument::where('temp_data_id', $del_temp_id)->delete();
-        ArtistTempDocument::where('temp_data_id', $del_temp_id)->update(['status' => 1]);
-        return redirect('company/artist/view_draft_details/' . $permit_id);
+        try {
+            DB::beginTransaction();
+            $del_temp_id  = $request->del_temp_id;
+            $permit_id = $request->del_permit_id;
+            ArtistTempData::where('id', $del_temp_id)->update(['del_status' => 1]);
+            ArtistTempDocument::where('temp_data_id', $del_temp_id)->update(['status' => 1]);
+            DB::commit();
+            $result = ['success', __('Artist Deleted Successfully'), 'Success'];
+        } catch (Exception $e) {
+            DB::rollBack();
+            $result = ['error', __($e->getMessage()), 'Error'];
+        }
+
+        return redirect(URL::signedRoute('company.view_draft_details', [ 'id' => $permit_id]))->with('message', $result);
     }
 
 
@@ -1313,8 +1370,21 @@ class ArtistController extends Controller
 
     public function update_artist_temp(Request $request)
     {
+        $toURL = '';
         try {
             DB::beginTransaction();
+
+            if($request->from == 'draft'){
+                $toURL = URL::signedRoute('company.view_draft_details', [ 'id' => $request->permit_id]);
+            }else if($request->from == 'amend'){
+                $toURL = URL::signedRoute('artist.permit', [ 'id' => $request->permit_id , 'from' => 'amend']);
+            }else if($request->from == 'edit') {
+                $toURL = URL::signedRoute('artist.permit', [ 'id' => $request->permit_id , 'from' => 'edit']);
+            }else if($request->from == 'renew') {
+                $toURL = URL::signedRoute('artist.permit', [ 'id' => $request->permit_id , 'from' => 'renew']);
+            }else if($request->from == 'new') {
+                $toURL = URL::signedRoute('company.add_new_permit', [ 'id' => $request->permit_id]);
+            }
 
         $temp_id = $request->temp_id;
         $artistDetails = json_decode($request->artistD, true);
@@ -1480,14 +1550,14 @@ class ArtistController extends Controller
         //     $result = ['error', __('Error, Please Try Again'), 'Error'];
         // }
 
-        return response()->json(['message' => $result]);
+        return response()->json(['message' => $result, 'toURL' => $toURL]);
     }
 
     public function permit(Request $request, $id, $status)
     {
-        // if(!$request->hasValidSignature()){
-        //     return abort(401);
-        // }
+        if(!$request->hasValidSignature()){
+            return abort(401);
+        }
         $permit_details = Permit::with('artistPermit', 'artistPermit.artist', 'artistPermit.profession', 'event')->where('created_by', Auth::user()->user_id)->where('permit_id', $id)->first();
 
         if (empty($permit_details)) {
@@ -1620,7 +1690,7 @@ class ArtistController extends Controller
         // dd($data_bundle['staff_comments']);
         $routeTo = '';
         if ($status == 'event') {
-            return redirect()->route('event.add_artist', $id);
+            return redirect(URL::signedRoute('event.add_artist', ['id' => $id]));
         }
         if ($status == 'edit') {
             $routeTo = 'permits.artist.edit_permit';
@@ -1640,18 +1710,15 @@ class ArtistController extends Controller
 
     public function edit_artist(Request $request , $id, $from)
     {
-        // if(!$request->hasValidSignature()){
-        //     return abort(401);
-        // }
+        if(!$request->hasValidSignature()){
+            return abort(401);
+        }
         $permit_id = ArtistTempData::where('id', $id)->value('permit_id');
         $artist_permit_id = ArtistTempData::where('id', $id)->value('artist_permit_id');
         $data_bundle = $this->preLoadData();
         $data_bundle['artist_details'] = ArtistTempData::with('Nationality', 'Profession')->where('id', $id)->first();
         $data_bundle['staff_comments'] = $artist_permit_id ? ArtistPermit::with('comments')->where('artist_permit_id', $artist_permit_id)->latest()->first() : '';
         $data_bundle['permit_details'] = ArtistPermit::with('artist', 'permit', 'artistPermitDocument', 'profession')->where('permit_id', $permit_id)->first();
-        if(is_null($data_bundle['artist_details'])){
-            return abort(401);
-        }
         $data_bundle['from'] = $from;
         $url = '';
         if ($from == 'amend') {
@@ -1667,8 +1734,10 @@ class ArtistController extends Controller
         ArtistPermitCheck::where('artist_permit_id', $id)->update(['status' => 1]);
     }
 
-    public function fetch_drafts()
+    public function fetch_drafts(Request $request)
     {
+        if($request->ajax())
+        {
         $user_id = Auth::user()->user_id;
         $drafts = ArtistTempData::with('profession')->where([
             ['status', 5],
@@ -1698,10 +1767,11 @@ class ArtistController extends Controller
             if(check_is_blocked()['status'] == 'blocked'){
                 return ;
             }
-            return '<a href="' . route('company.view_draft_details', $permit->permit_id) . '"><span class="kt-badge kt-badge--warning kt-badge--inline">'.__('View').'</span></a>&emsp;<span onClick="delete_draft(' . $permit->permit_id . ')" data-toggle="modal"  class="kt-badge kt-badge--danger kt-badge--inline">'.__('Remove').'</span>';
+            return '<a href="' . \Illuminate\Support\Facades\URL::signedRoute('company.view_draft_details', $permit->permit_id) . '"><span class="kt-badge kt-badge--warning kt-badge--inline">'.__('View').'</span></a>&emsp;<span onClick="delete_draft(' . $permit->permit_id . ')" data-toggle="modal"  class="kt-badge kt-badge--danger kt-badge--inline">'.__('Remove').'</span>';
         })->addColumn('details', function ($permit) {
             return '<a href="' . \Illuminate\Support\Facades\URL::signedRoute('company.get_draft_details', $permit->permit_id) . '" title="View Details" class="kt-font-dark"><i class="fa fa-file"></i></a>';
         })->rawColumns(['action', 'details'])->make(true);
+    }
     }
 
     public function get_draft_details(Request $request, $permit_id)
@@ -1734,14 +1804,14 @@ class ArtistController extends Controller
             DB::rollBack();
             $result = ['error', __($e->getMessage()), 'Error'];
         }
-        return redirect()->route('artist.index')->with('message', $result);
+        return redirect(URL::signedRoute('artist.index').'#draft')->with('message', $result);
     }
 
     public function view_draft_details(Request $request, $id)
     {
-        // if(!$request->hasValidSignature()){
-        //     return abort(401);
-        // }
+        if(!$request->hasValidSignature()){
+            return abort(401);
+        }
 
         $user_id = Auth::user()->user_id;
         $data['artist_details'] = ArtistTempData::with('profession', 'nationality', 'ArtistTempDocument', 'event')->where([
@@ -1758,6 +1828,12 @@ class ArtistController extends Controller
 
     public function add_draft(Request $request)
     {
+        $toURL = URL::signedRoute('artist.index').'#draft';
+        try{ 
+
+            DB::beginTransaction();
+
+
         $temp_permit_id = $request->temp_permit_id;
         $user_id = Auth::user()->user_id;
 
@@ -1784,14 +1860,20 @@ class ArtistController extends Controller
             ['del_status', 1],
         ])->delete();
 
-        if ($update) {
             $this->makeSessionForgetPermitDetails();
-            $result = ['success', __('Draft Saved Successfully'), 'Success'];
-        } else {
-            $result = ['error', __('Error, Please Try Again'), 'Error'];
-        }
+        
 
-        return response()->json(['message' => $result]);
+        DB::commit();
+		
+        $result = ['success', __('Draft Saved Successfully'), 'Success'];
+	} catch (Exception $e) {
+		DB::rollBack();
+		
+		$result = ['error', __($e->getMessage()), 'Error'];
+	}
+	
+
+        return response()->json(['message' => $result , 'toURL' => $toURL]);
     }
 
     public function edit_artist_draft(Request $request, $temp_id)
@@ -1810,6 +1892,7 @@ class ArtistController extends Controller
 
     public function update_permit(Request $request)
     {
+        $toURL = URL::signedRoute('artist.index').'#applied';
         try {
             DB::beginTransaction();
 
@@ -2035,7 +2118,7 @@ class ArtistController extends Controller
             $result = ['error', __($e->getMessage()), 'Error'];
         }
 
-        return response()->json(['message' => $result]);
+        return response()->json(['message' => $result , 'toURL' => $toURL]);
     }
 
     public function get_temp_photo_temp_id($id)
@@ -2130,9 +2213,11 @@ class ArtistController extends Controller
         $transaction_id = $request->transactionId;
         $receipt = $request->receipt; 
         $order_id = $request->orderId;
+        $toURL = '';
         try {
             DB::beginTransaction();
-
+            $toURL = URL::signedRoute('company.happiness_center', [ 'id' => $permit_id]);
+            
         $transArr = Transaction::create([
             'reference_number' => getTransactionReferNumber(),
             'transaction_type' => 'artist',
@@ -2269,19 +2354,20 @@ class ArtistController extends Controller
         }
 
 
-        return response()->json(['message' => $result]);
+        return response()->json(['message' => $result , 'toURL' => $toURL]);
     }
  
     public function happiness_center(Request $request, $id)
     {
-        // if(!$request->hasValidSignature()){
-        //     return abort(401);
-        // }
+        if(!$request->hasValidSignature()){
+            return abort(401);
+        }
         return view('permits.artist.happinessmeter', ['id' => $id]);
     }
 
     public function submit_happiness(Request $request)
     {
+        $toURL = URL::signedRoute('artist.index').'#applied';
         try {
             DB::beginTransaction();
 
@@ -2351,7 +2437,7 @@ class ArtistController extends Controller
             $result = ['error', __($e->getMessage()), 'Error'];
         }
 
-        return response()->json(['message' => $result]);
+        return response()->json(['message' => $result, 'toURL' => $toURL]);
     }
 
     public function checkVisaRequired($id)
